pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.4",
    "aiosqlite>=0.21.0",
    "asyncio>=3.4.3",
    "discord-py>=2.5.2",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "youtube-dl>=2021.12.17",
]
